move impact mocks to separate file, add impact controller tests
